refactor sstable index
